Refactor TestRunProxy into its own file, clean up docs
`TestRunProxy` was colocated in the same file as `TestRunner`. Move it
to its own file and clean up its documentation.
